# make multiplication of polynomials type stable